Move PublicApiSteps to the utils package for using it in REST and UI tests
